+++ CHANGES.rst
@@ -1,6 +1,25 @@
 Version history
 ===============
 
+**UNRELEASED**
+
+- Type annotations for ARRAY column attributes now include the Python type of
+  the array elements
+- Added support for specifying engine arguments via ``--engine-arg``
+  (PR by @LajosCseppento)
+
+**3.0.0**
+
+- Dropped support for Python 3.8
+- Changed nullable relationships to include ``Optional`` in their type annotations
+- Fixed SQLModel code generation
+- Fixed two rendering issues in ``ENUM`` columns when a non-default schema is used: an
+  unwarranted positional argument and missing the ``schema`` argument
+- Fixed ``AttributeError`` when metadata contains user defined column types
+- Fixed ``AssertionError`` when metadata contains a column type that is a type decorator
+  with an all-uppercase name
+- Fixed MySQL ``DOUBLE`` column types being rendered with the wrong arguments
+
 **3.0.0rc5**
 
 - Fixed pgvector support not working